osv: fix one compile issue
make[1]: Entering directory `/home/zwu/virt/osv/build/release'
  GEN gen/include/bits/alltypes.h
  AS arch/x64/boot16.o
  LD boot.bin
  AS arch/x64/boot.o
  CXX loader.o
In file included from <command-line>:0:0:
./../../compiler/include/intrinsics.hh:10:127: error: new declaration ‘unsigned char __builtin_ia32_addcarryx_u64(unsigned char, long unsigned int, long unsigned int, long long unsigned int*)’ [-Werror]
./../../compiler/include/intrinsics.hh:10:127: error: ambiguates built-in declaration ‘unsigned char __builtin_ia32_addcarryx_u64(unsigned char, long long unsigned int, long long unsigned int, long long unsigned int*)’ [-Werror]
cc1plus: all warnings being treated as errors
make[1]: *** [loader.o] Error 1
make[1]: Leaving directory `/home/zwu/virt/osv/build/release'
make: *** [all] Error 2

+++ compiler/include/intrinsics.hh
@@ -7,7 +7,7 @@
 #if __GNUC__ == 4 && __GNUC_MINOR__ < 8
 
 extern unsigned char __builtin_ia32_addcarryx_u32(unsigned char, unsigned int, unsigned int, unsigned int*);
-extern unsigned char __builtin_ia32_addcarryx_u64(unsigned char, long unsigned int, long unsigned int, long long unsigned int*);
+extern unsigned char __builtin_ia32_addcarryx_u64(unsigned char, long long unsigned int, long long unsigned int, long long unsigned int*);
 
 #endif
